Fix blending in map mode (#282)
When map mode is enabled, the scene is rendered thrice to a single
frame. This breaks blending in GL renderer, which depends on a clear
depth buffer.

Don't wait for game loop to swap the frame buffers and manually clear
the depth buffer as soon as the scene is done rendering.

Signed-off-by: Artur Rojek <contact@artur-rojek.eu>