Fix sound sample mixing (#281)
The sound engine used in Dethrace is unable to play the same sample more
than once at a time. If two or more sound sources request to play the
same sound id, the one to use it last will overwrite any previous
effects on the sample. This is most prominent for car engine noises and
cop sirens.
This issue also occurs in Windows builds of OG, but not in DOS ones.

Fix this issue by duplicating sound samples and storing a copy
per-channel.

Signed-off-by: Artur Rojek <contact@artur-rojek.eu>